# Conflicts:
#   backend/stockolm/src/main/java/com/example/stockolm/domain/analystBoard/repository/AnalystBoardCustomRepository.java
#   backend/stockolm/src/main/java/com/example/stockolm/domain/analystBoard/repository/AnalystBoardCustomRepositoryImpl.java

+++ backend/stockolm/src/main/java/com/example/stockolm/domain/follow/dto/response/FollowAnalystResponse.java
@@ -8,12 +8,12 @@ public class FollowAnalystResponse {
     private String userName;
     private String userNickName;
     private String userImagePath;
-    private float reliability;
-    private float accuracy;
+    private Double reliability;
+    private Double accuracy;
     private int totalAnalystRanking;
 
     @Builder
-    public FollowAnalystResponse(String userName, String userNickName, String userImagePath, float reliability, float accuracy, int totalAnalystRanking) {
+    public FollowAnalystResponse(String userName, String userNickName, String userImagePath, Double reliability, Double accuracy, int totalAnalystRanking) {
         this.userName = userName;
         this.userNickName = userNickName;
         this.userImagePath = userImagePath;
